start refactoring dom/string renderers
